update xxh3 to 0.8.0, fix is_ascii
